backend/requirements.txt
# FastAPI 核心
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 数据库
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
alembic==1.13.1

# 数据校验
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP 客户端
httpx==0.26.0
aiohttp==3.9.3

# AI 服务
dashscope==1.14.1

# 缓存
redis==5.0.1
aioredis==2.0.1

# 工具
python-dotenv==1.0.1
Pillow==10.2.0
python-dateutil==2.8.2

# 开发工具
pytest==8.0.0
pytest-asyncio==0.23.4


